requirements.txt
djangorestframework==3.14.0
Django==3.2.5
django-crispy-forms==2.0
crispy-forms-gds==0.2.4
Pillow==9.4.0
requests==2.28.2
requests-oauthlib==1.3.1
aiogram==2.25.1
